# perf: improve runtime performance