--- a/.changeset/spicy-pears-love.md
+++ b/.changeset/spicy-pears-love.md
@@ -1,0 +1,7 @@
+---
+'@pandacss/generator': patch
+'@pandacss/shared': patch
+---
+
+Improve the performance of the runtime transform functions by caching their results (css, cva, sva, recipe/slot recipe,
+patterns)