--- a/packages/generator/src/artifacts/js/recipe.ts
+++ b/packages/generator/src/artifacts/js/recipe.ts
@@ -27,7 +27,7 @@ export function generateCreateRecipe(ctx: Context) {
    ${ctx.file.import('css', '../css/css')}
    ${ctx.file.import('assertCompoundVariant, getCompoundVariantCss', '../css/cva')}
    ${ctx.file.import('cx', '../css/cx')}
-   ${ctx.file.import('compact, createCss, memo, splitProps, uniq, withoutSpace', '../helpers')}
+   ${ctx.file.import('compact, createCss, splitProps, uniq, withoutSpace', '../helpers')}
 
    export const createRecipe = (name, defaultVariants, compoundVariants) => {
     const getRecipeStyles = (variants) => {
@@ -73,7 +73,7 @@ export function generateCreateRecipe(ctx: Context) {
       return recipeCss(recipeStyles)
      }
 
-      return Object.assign(memo(recipeFn), {
+      return Object.assign(recipeFn, {
         __getCompoundVariantCss__: (variants) => {
           return getCompoundVariantCss(compoundVariants, getRecipeStyles(variants));
         },
@@ -122,7 +122,7 @@ export function generateRecipes(ctx: Context, filters?: ArtifactFilters) {
       .when(
         Recipes.isSlotRecipeConfig,
         (config) => outdent`
-        ${ctx.file.import('splitProps, getSlotCompoundVariant', '../helpers')}
+        ${ctx.file.import('getSlotCompoundVariant, memo, splitProps', '../helpers')}
         ${ctx.file.import('createRecipe', './create-recipe')}
 
         const ${baseName}DefaultVariants = ${stringify(defaultVariants ?? {})}
@@ -131,9 +131,9 @@ export function generateRecipes(ctx: Context, filters?: ArtifactFilters) {
         const ${baseName}SlotNames = ${stringify(config.slots.map((slot) => [slot, `${config.className}__${slot}`]))}
         const ${baseName}SlotFns = /* @__PURE__ */ ${baseName}SlotNames.map(([slotName, slotKey]) => [slotName, createRecipe(slotKey, ${baseName}DefaultVariants, getSlotCompoundVariant(${baseName}CompoundVariants, slotName))])
 
-        const ${baseName}Fn = (props = {}) => {
+        const ${baseName}Fn = memo((props = {}) => {
           return Object.fromEntries(${baseName}SlotFns.map(([slotName, slotFn]) => [slotName, slotFn(props)]))
-        }
+        })
 
         const ${baseName}VariantKeys = ${stringify(Object.keys(variantKeyMap))}
 
@@ -151,7 +151,7 @@ export function generateRecipes(ctx: Context, filters?: ArtifactFilters) {
       )
       .otherwise(
         (config) => outdent`
-        ${ctx.file.import('splitProps', '../helpers')}
+        ${ctx.file.import('memo, splitProps', '../helpers')}
         ${ctx.file.import('createRecipe, mergeRecipes', './create-recipe')}
 
         const ${baseName}Fn = /* @__PURE__ */ createRecipe('${config.className}', ${stringify(
@@ -162,7 +162,7 @@ export function generateRecipes(ctx: Context, filters?: ArtifactFilters) {
 
         const ${baseName}VariantKeys = Object.keys(${baseName}VariantMap)
 
-        export const ${baseName} = /* @__PURE__ */ Object.assign(${baseName}Fn, {
+        export const ${baseName} = /* @__PURE__ */ Object.assign(memo(${baseName}Fn), {
           __recipe__: true,
           __name__: '${baseName}',
           raw: (props) => props,