--- a/sandbox/codegen/__tests__/frameworks/qwik.styled-factory.test.tsx
+++ b/sandbox/codegen/__tests__/frameworks/qwik.styled-factory.test.tsx
@@ -250,7 +250,7 @@ describe('styled factory - button recipe', async () => {
 
     const container = screen.querySelector('button')!
     expect(container.outerHTML).toMatchInlineSnapshot(
-      '"<button class=\\"button button--visual_outline button--size_md text_red.200 mx_2 fs_xl custom-btn\\">Click me</button>"',
+      '"<button class=\\"button button--visual_outline button--size_md mx_2 text_red.200 fs_xl custom-btn\\">Click me</button>"',
     )
   })