--- a/sandbox/codegen/__tests__/frameworks/solid.styled-factory.test.tsx
+++ b/sandbox/codegen/__tests__/frameworks/solid.styled-factory.test.tsx
@@ -298,7 +298,7 @@ describe('styled factory - button recipe', () => {
 
     expect(container.firstChild).toMatchInlineSnapshot(`
       <button
-        class="button button--size_md button--visual_outline text_red.200 mx_2 fs_xl custom-btn"
+        class="button button--size_md button--visual_outline mx_2 text_red.200 fs_xl custom-btn"
       >
         Click me
       </button>