--- a/sandbox/codegen/__tests__/frameworks/vue.styled-factory.test.tsx
+++ b/sandbox/codegen/__tests__/frameworks/vue.styled-factory.test.tsx
@@ -296,7 +296,7 @@ describe('styled factory - button recipe', () => {
     const { firstChild } = container as HTMLElement
     expect(firstChild).toMatchInlineSnapshot(`
       <button
-        class="button button--visual_outline button--size_md text_red.200 mx_2 fs_xl custom-btn"
+        class="button button--visual_outline button--size_md mx_2 text_red.200 fs_xl custom-btn"
       >
         Click me
       </button>