--- a/sandbox/runtime-perf/postcss.config.cjs
+++ b/sandbox/runtime-perf/postcss.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  plugins: {
+    '@pandacss/dev/postcss': {},
+    autoprefixer: {},
+  },
+}