--- a/sandbox/runtime-perf/src/index.css
+++ b/sandbox/runtime-perf/src/index.css
@@ -1,0 +1,1 @@
+@layer reset, base, tokens, recipes, utilities;