--- a/sandbox/runtime-perf/vite.config.ts
+++ b/sandbox/runtime-perf/vite.config.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+
+const ANALYZE = !!process.env.ANALYZE
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react()],
+  build: {
+    sourcemap: ANALYZE,
+  },
+  resolve: {
+    conditions: ['source'],
+  },
+})