… packages/generator/src/artifacts/generated/helpers.mjs.json
… pnpm-lock.yaml
… sandbox/runtime-perf/panda.config.ts