Closes https://github.com/chakra-ui/panda/discussions/1982

## 📝 Description

Improve the performance of the runtime transform functions by caching their results (css, cva, sva, recipe/slot recipe,
patterns)

since Panda has to transform style objects to classes (`{ backgroundColor: "red.300" }` -> `bg_red`, it will always have a little bit of overhead vs just writing the final classes, caching those reduce the time spent on it.

## 💣 Is this a breaking change (Yes/No):

no

## 📝 Additional Information

if you're using anything that can pre-render to HTML (Astro, RSC, etc) you can avoid this issue entirely by automatically inlining classes at build-time, without ever needing those runtime functions