@segunadebayo 2024-01-11
Great work! Did some performance comparisons, rendering 100 instances of a component using cva.

**Before**
<img width="1728" alt="CleanShot 2024-01-11 at 15 45 21@2x" src="https://github.com/chakra-ui/panda/assets/6916170/3966bb16-9e69-4177-9ca0-4189525e3683">

**After**
<img width="1726" alt="CleanShot 2024-01-11 at 15 45 39@2x" src="https://github.com/chakra-ui/panda/assets/6916170/714f0088-3ca9-4168-a3e0-a69e5dba6596">


**Diff**
Mount Time: 10.9s -> 3.6s (100.69% improvement)
Update Time: 9s -> 4.8s (60.8696% difference)